# src/main/java/com/example/bankcards/controller/AuthController.java
package com.example.bankcards.controller;

import com.example.bankcards.dto.SignupResponseDto;
import com.example.bankcards.entity.dto.LoginRequestDto;
import com.example.bankcards.entity.dto.LoginResponseDto;
import com.example.bankcards.security.AuthService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/auth")
@RequiredArgsConstructor
public class AuthController {
    private final AuthService authService;

    @PostMapping(path = "/login", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<LoginResponseDto> login(@RequestBody LoginRequestDto loginRequestDto) {
        return ResponseEntity.ok(authService.login(loginRequestDto));
    }

    @PostMapping(path = "/signup", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<SignupResponseDto> signup(@RequestBody LoginRequestDto loginRequestDto) {
        return ResponseEntity.ok(authService.signup(loginRequestDto));
    }
}